Title: 세 수의 합과 평균 구하기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241016/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.c
@@ -3,6 +3,6 @@
 int main() {
     int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
-    printf("%d %.0lf",a+b+c,(double(a+b+c)/3));
+    printf("%d %d",a+b+c,(a+b+c)/3);
     return 0;
 }
